remove legacy vertical tabs
