+++ frontend/src/scenes/insights/InsightTabs/RetentionTab.tsx
@@ -4,42 +4,36 @@ import { PropertyFilters } from 'lib/components/PropertyFilters/PropertyFilters'
 import { ActionFilterDropdown } from '../ActionFilter/ActionFilterRow/ActionFilterDropdown'
 import { entityFilterLogic } from '../ActionFilter/entityFilterLogic'
 
-import { DownOutlined, InfoCircleOutlined, ExportOutlined } from '@ant-design/icons'
+import { DownOutlined, InfoCircleOutlined } from '@ant-design/icons'
 import {
     retentionTableLogic,
     dateOptions,
-    retentionOptions,
     retentionOptionDescriptions,
+    defaultFilters,
 } from 'scenes/retention/retentionTableLogic'
-import { Button, Select, Tooltip } from 'antd'
-import { Link } from 'lib/components/Link'
-import { CloseButton } from 'lib/components/CloseButton'
-import dayjs from 'dayjs'
-import dayjsGenerateConfig from 'rc-picker/lib/generate/dayjs'
-import generatePicker from 'antd/es/date-picker/generatePicker'
-import { FilterType } from '~/types'
+import { Button, Select, Tooltip, Row, Col, Skeleton } from 'antd'
+
+import { FilterType, RetentionType } from '~/types'
 import { TestAccountFilter } from '../TestAccountFilter'
 import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
 import './RetentionTab.scss'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { RetentionTabHorizontal } from './RetentionTabHorizontal'
-import { FEATURE_FLAGS } from 'lib/constants'
+import { RETENTION_FIRST_TIME, RETENTION_RECURRING } from 'lib/constants'
+import useBreakpoint from 'antd/lib/grid/hooks/useBreakpoint'
+import { IconExternalLink } from 'lib/components/icons'
 import { BaseTabProps } from '../Insights'
+import { InsightTitle } from './InsightTitle'
+import { InsightActionBar } from './InsightActionBar'
 
-const DatePicker = generatePicker<dayjs.Dayjs>(dayjsGenerateConfig)
-
-export function RetentionTab(props: BaseTabProps): JSX.Element {
-    const { featureFlags } = useValues(featureFlagLogic)
-    return featureFlags[FEATURE_FLAGS.QUERY_UX_V2] ? <RetentionTabHorizontal {...props} /> : <DefaultRetentionTab />
-}
-
-function DefaultRetentionTab(): JSX.Element {
+export function RetentionTab({ annotationsToCreate }: BaseTabProps): JSX.Element {
     const node = useRef<HTMLElement>(null)
     const returningNode = useRef<HTMLElement>(null)
     const [open, setOpen] = useState<boolean>(false)
     const [returningOpen, setReturningOpen] = useState<boolean>(false)
-    const { filters, actionsLookup } = useValues(retentionTableLogic({ dashboardItemId: null }))
+    const { filters, actionsLookup, filtersLoading } = useValues(retentionTableLogic({ dashboardItemId: null }))
     const { setFilters } = useActions(retentionTableLogic({ dashboardItemId: null }))
+
+    const screens = useBreakpoint()
+    const isSmallScreen = screens.xs || (screens.sm && !screens.md)
 
     const entityLogic = entityFilterLogic({
         setFilters: (newFilters: FilterType) => {
@@ -83,128 +77,129 @@ function DefaultRetentionTab(): JSX.Element {
         (filters.target_entity.id && actionsLookup[filters.target_entity.id]) ||
         'Select action'
 
+    // TODO: Update constant in retentionTableLogic.ts when releasing 4050
+    const retentionOptions = {
+        [`${RETENTION_FIRST_TIME}`]: 'for the first time',
+        [`${RETENTION_RECURRING}`]: 'recurringly',
+    }
+
     return (
         <div data-attr="retention-tab" className="retention-tab">
-            <h4 className="secondary">
-                Cohortizing Event
-                <Tooltip
-                    key="2"
-                    placement="right"
-                    title={`Event that determines which users are considered to form each cohort (i.e. performed event in ${filters.period} 0)`}
-                >
-                    <InfoCircleOutlined className="info-indicator" />
-                </Tooltip>
-            </h4>
-            <div style={{ display: '-webkit-inline-box', flexWrap: 'wrap' }}>
-                <Button
-                    ref={node}
-                    data-attr="retention-action"
-                    onClick={(): void => setOpen(!open)}
-                    style={{ marginRight: 8, marginBottom: 8 }}
-                    className="btn-retention-dropdown"
-                >
-                    <PropertyKeyInfo value={selectedCohortizingEvent} />
-                    <DownOutlined className="dropdown-indicator" />
-                </Button>
-                <Select
-                    value={retentionOptions[filters.retention_type]}
-                    onChange={(value): void => setFilters({ retention_type: value })}
-                    dropdownMatchSelectWidth={false}
-                >
-                    {Object.entries(retentionOptions).map(([key, value]) => (
-                        <Select.Option key={key} value={key}>
-                            {value}
-                            <Tooltip placement="right" title={retentionOptionDescriptions[key]}>
-                                <InfoCircleOutlined className="info-indicator" />
-                            </Tooltip>
-                        </Select.Option>
-                    ))}
-                </Select>
-                <ActionFilterDropdown
-                    open={open}
-                    logic={entityLogic as any}
-                    openButtonRef={node}
-                    onClose={() => setOpen(false)}
-                />
-            </div>
-            <h4 style={{ marginTop: '0.5rem' }} className="secondary">
-                Retaining event
-                <Tooltip
-                    key="3"
-                    placement="right"
-                    title="Event that determines if each user came back on each period (i.e. if they were retained)"
-                >
-                    <InfoCircleOutlined className="info-indicator" />
-                </Tooltip>
-            </h4>
-
-            <Button
-                ref={returningNode}
-                data-attr="retention-returning-action"
-                onClick={(): void => setReturningOpen(!returningOpen)}
-                className="btn-retention-dropdown"
-            >
-                <PropertyKeyInfo value={selectedRetainingEvent} />
-                <DownOutlined className="dropdown-indicator" />
-            </Button>
-            <ActionFilterDropdown
-                open={returningOpen}
-                logic={entityLogicReturning as any}
-                openButtonRef={returningNode}
-                onClose={() => setReturningOpen(false)}
-            />
-            <div className="mt-05">
-                <Link
-                    to="https://posthog.com/docs/features/retention?utm_campaign=learn-more&utm_medium=in-product"
-                    target="_blank"
-                    rel="noreferrer noopener"
-                >
-                    More info on retention <ExportOutlined />
-                </Link>
-            </div>
-
-            <hr />
-            <h4 className="secondary">Filters</h4>
-            <PropertyFilters pageKey="insight-retention" />
-            <TestAccountFilter filters={filters} onChange={setFilters} />
-            <>
-                <hr />
-                <h4 className="secondary">Current Date</h4>
-                <div>
-                    <DatePicker
-                        showTime={filters.period === 'Hour'}
-                        use12Hours
-                        format={filters.period === 'Hour' ? 'YYYY-MM-DD, h a' : 'YYYY-MM-DD'}
-                        className="mb-05"
-                        value={filters.date_to && dayjs(filters.date_to)}
-                        onChange={(date_to): void => setFilters({ date_to: date_to && dayjs(date_to).toISOString() })}
-                        allowClear={false}
+            <Row gutter={16}>
+                <Col md={16} xs={24}>
+                    <InsightTitle
+                        actionBar={
+                            <InsightActionBar
+                                filters={filters}
+                                annotations={annotationsToCreate}
+                                insight="RETENTION"
+                                onReset={() => setFilters(defaultFilters({}))}
+                            />
+                        }
                     />
-                    {filters.date_to && (
-                        <CloseButton
-                            onClick={() => setFilters({ date_to: null })}
-                            style={{
-                                marginLeft: 8,
-                            }}
-                        />
+                    <Row gutter={8} align="middle">
+                        <Col>
+                            Showing <b>Unique users</b> who did
+                        </Col>
+                        <Col>
+                            <Button
+                                className="btn-retention-dropdown"
+                                ref={node}
+                                data-attr="retention-action"
+                                onClick={() => setOpen(!open)}
+                            >
+                                <PropertyKeyInfo value={selectedCohortizingEvent} disablePopover />
+                                <DownOutlined className="dropdown-indicator" />
+                            </Button>
+                            <ActionFilterDropdown
+                                open={open}
+                                logic={entityLogic as any}
+                                openButtonRef={node}
+                                onClose={() => setOpen(false)}
+                            />
+                        </Col>
+                        <Col>
+                            <div style={{ display: '-webkit-inline-box', flexWrap: 'wrap' }}>
+                                <Select
+                                    value={retentionOptions[filters.retention_type]}
+                                    onChange={(value): void => setFilters({ retention_type: value as RetentionType })}
+                                    dropdownMatchSelectWidth={false}
+                                >
+                                    {Object.entries(retentionOptions).map(([key, value]) => (
+                                        <Select.Option key={key} value={key}>
+                                            {value}
+                                            <Tooltip placement="right" title={retentionOptionDescriptions[key]}>
+                                                <InfoCircleOutlined className="info-indicator" />
+                                            </Tooltip>
+                                        </Select.Option>
+                                    ))}
+                                </Select>
+                            </div>
+                        </Col>
+                        <Col>grouped by</Col>
+                        <Col>
+                            <Select
+                                value={filters.period}
+                                onChange={(value): void => setFilters({ period: value })}
+                                dropdownMatchSelectWidth={false}
+                            >
+                                {dateOptions.map((period) => (
+                                    <Select.Option key={period} value={period}>
+                                        {period}
+                                    </Select.Option>
+                                ))}
+                            </Select>
+                        </Col>
+                    </Row>
+                    <Row gutter={8} align="middle" className="mt">
+                        <Col>... who then came back and did</Col>
+                        <Col>
+                            <Button
+                                ref={returningNode}
+                                data-attr="retention-returning-action"
+                                onClick={(): void => setReturningOpen(!returningOpen)}
+                                className="btn-retention-dropdown"
+                            >
+                                <PropertyKeyInfo value={selectedRetainingEvent} disablePopover />
+                                <DownOutlined className="dropdown-indicator" />
+                            </Button>
+                            <ActionFilterDropdown
+                                open={returningOpen}
+                                logic={entityLogicReturning as any}
+                                openButtonRef={returningNode}
+                                onClose={() => setReturningOpen(false)}
+                            />
+                        </Col>
+                    </Row>
+                    <Row>
+                        <Col>
+                            <p className="text-muted mt">
+                                Want to learn more about retention?{' '}
+                                <a
+                                    href="https://posthog.com/docs/features/retention?utm_campaign=learn-more-horizontal&utm_medium=in-product"
+                                    target="_blank"
+                                    rel="noopener"
+                                    style={{ display: 'inline-flex', alignItems: 'center' }}
+                                >
+                                    Go to docs
+                                    <IconExternalLink style={{ marginLeft: 4 }} />
+                                </a>
+                            </p>
+                        </Col>
+                    </Row>
+                </Col>
+                <Col md={8} xs={24} style={{ marginTop: isSmallScreen ? '2rem' : 0 }}>
+                    <h4 className="secondary">Global Filters</h4>
+                    {filtersLoading ? (
+                        <Skeleton active paragraph={{ rows: 1 }} />
+                    ) : (
+                        <>
+                            <PropertyFilters pageKey="insight-retention" />
+                            <TestAccountFilter filters={filters} onChange={setFilters} />
+                        </>
                     )}
-                </div>
-                <hr />
-                <h4 className="secondary">Period</h4>
-                <div>
-                    <Select
-                        value={filters.period}
-                        onChange={(value): void => setFilters({ period: value })}
-                        dropdownMatchSelectWidth={false}
-                    >
-                        {dateOptions.map((period) => (
-                            <Select.Option key={period} value={period}>
-                                {period}
-                            </Select.Option>
-                        ))}
-                    </Select>
-                </div>
-            </>
+                </Col>
+            </Row>
         </div>
     )
 }
